Update level-up and rename to explore
