+++ datumaro/cli/commands/__init__.py
@@ -15,7 +15,7 @@ from . import (
     info,
     merge,
     patch,
-    search,
+    explore,
     stats,
     transform,
     validate,
@@ -40,7 +40,7 @@ def get_non_project_commands():
         ("generate", generate, "Generate synthetic dataset"),
         ("merge", merge, "Merge datasets"),
         ("patch", patch, "Update dataset from another one"),
-        ("search", search, "Search similar datasetitems of query"),
+        ("explore", explore, "Explore similar datasetitems of query"),
         ("stats", stats, "Compute dataset statistics"),
         ("transform", transform, "Modify dataset items"),
         ("validate", validate, "Validate dataset"),
