+++ docs/source/docs/reference/cli/commands/datumaro.cli.commands.explore.rst
@@ -1,0 +1,7 @@
+Explore module
+=============
+
+.. automodule:: datumaro.cli.commands.explore
+   :members:
+   :undoc-members:
+   :show-inheritance:
