--- docs/source/docs/reference/cli/commands/datumaro.cli.commands.search.rst
@@ -1,7 +1,0 @@
-Search module
-=============
-
-.. automodule:: datumaro.cli.commands.search
-   :members:
-   :undoc-members:
-   :show-inheritance:
