+++ docs/source/docs/reference/components/components/datumaro.components.explorer.rst
@@ -1,0 +1,7 @@
+Explorer module
+===============
+
+.. automodule:: datumaro.components.explorer
+   :members:
+   :undoc-members:
+   :show-inheritance:
