--- docs/source/docs/reference/components/components/datumaro.components.searcher.rst
@@ -1,7 +1,0 @@
-Searcher module
-===============
-
-.. automodule:: datumaro.components.searcher
-   :members:
-   :undoc-members:
-   :show-inheritance:
