+++ docs/source/docs/reference/plugins/plugins/datumaro.plugins.explorer.rst
@@ -1,0 +1,7 @@
+Explorer module
+===============
+
+.. automodule:: datumaro.plugins.explorer
+   :members:
+   :undoc-members:
+   :show-inheritance:
