--- docs/source/docs/reference/plugins/plugins/datumaro.plugins.searcher.rst
@@ -1,7 +1,0 @@
-Searcher module
-===============
-
-.. automodule:: datumaro.plugins.searcher
-   :members:
-   :undoc-members:
-   :show-inheritance:
